+++ src/components/Header.vue
@@ -3,12 +3,24 @@
 </script>
 
 <template>
-  <nav>
-    <ul style="display: flex; list-style: none; gap: 10px; margin: 0; padding: 0;">
-      <li>a</li>
-      <li>b</li>
-      <li>c</li>
-    </ul>
-  </nav>
+  <a class="logo" href="/portfolio">
+    BS
+  </a>
   <DarkMode />
 </template>
+
+<style>
+  #header .logo {
+    text-decoration: none;
+    font-size: 2rem;
+    font-weight: 700;
+  }
+
+  #header nav ul {
+    display: flex;
+    list-style: none;
+    gap: 10px;
+    margin: 0;
+    padding: 0;
+  }
+</style>