+++ src/components/Main.vue
@@ -4,7 +4,7 @@
     <p>I'm a Senior Frontend Web Developer, and this is where I plan to host some personal projects.</p>
     <h4>Coming Soon</h4>
     <ol>
-      <li>Dark Mode Toggle</li>
+      <li><span style="text-decoration: line-through">Dark Mode Toggle</span></li>
       <li>SSFP (Searchable, Sortable, Filterable, Paginated) Datatable</li>
       <li>D3 Interactive Charts &amp; Maps</li>
       <li>Various CSS Artboards...</li>